Actualización del modelo de BD y creación de la jerarquía de usuarios.
