+++ src/main/java/com/uis/sisgan/persistence/FarmerRepository.java
@@ -2,7 +2,7 @@ package com.uis.sisgan.persistence;
 
 import com.uis.sisgan.persistence.crud.FarmerCrudRepository;
 import com.uis.sisgan.persistence.entity.Cattle;
-import com.uis.sisgan.persistence.entity.Farmer;
+import com.uis.sisgan.persistence.entity.Propietary;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -19,8 +19,8 @@ public class FarmerRepository  {
         y que retornan. este es el verdadero Repository
 
      */
-    public List<Farmer> getAll(){
-        return (List<Farmer>) farmerCrudRepository.findAll();
+    public List<Propietary> getAll(){
+        return (List<Propietary>) farmerCrudRepository.findAll();
     }
